Refactored reply action out of handlers.ts
ref https://linear.app/ghost/issue/PROD-669

In order to start restricting the reply action when a user is blocked we want
to separate it out of the handlers file and start following the pattern of
controller -> service -> domain
